# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateDatabaseRequest": {
        "properties": {
          "name": {
            "maxLength": 64,
            "minLength": 2,
            "pattern": "^[a-z][a-z0-9_]{1,63}$",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateDatabaseRequest",
        "type": "object"
      },
      "CreateSiteRequest": {
        "properties": {
          "create_dns_zone": {
            "default": false,
            "title": "Create Dns Zone",
            "type": "boolean"
          },
          "domain": {
            "maxLength": 253,
            "minLength": 1,
            "title": "Domain",
            "type": "string"
          },
          "php_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Php Version"
          }
        },
        "required": [
          "domain"
        ],
        "title": "CreateSiteRequest",
        "type": "object"
      },
...
      "CredentialsResponse": {
        "properties": {
          "database": {
            "$ref": "#/components/schemas/DatabaseResponse"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "database",
          "password"
        ],
        "title": "CredentialsResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperationAccepted": {
        "properties": {
          "operation_id": {
            "title": "Operation Id",
            "type": "integer"
          },
          "site": {
            "$ref": "#/components/schemas/SiteResponse"
          }
        },
        "required": [
          "site",
          "operation_id"
        ],
        "title": "OperationAccepted",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "HostyPanel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/databases/sites/{site_id}": {
      "post": {
        "operationId": "create_database_api_databases_sites__site_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatabaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Database",
        "tags": [
          "databases"
        ]
      }
    },
...
    "/api/sites": {
      "get": {
        "operationId": "list_sites_api_sites_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SiteResponse"
                  },
                  "title": "Response List Sites Api Sites Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Sites",
        "tags": [
          "sites"
        ]
      },
      "post": {
        "operationId": "create_site_api_sites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSiteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Site",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/sites/{site_id}": {
      "delete": {
        "operationId": "delete_site_api_sites__site_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteSiteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Site",
        "tags": [
          "sites"
        ]
      },
      "get": {
        "operationId": "get_site_api_sites__site_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Site",
        "tags": [
          "sites"
        ]
      }
    },
...
    "/api/sites/{site_id}/cloudflare-proxy": {
      "patch": {
        "description": "Toggle 'behind the Cloudflare proxy' for a site.\n\nWhen enabled, Caddy issues an internal origin certificate for the domain\ninstead of attempting ACME HTTP-01 (which the proxy would break); Cloudflare\nterminates public TLS at the edge (SSL mode \"Full\").",
        "operationId": "set_cloudflare_proxy_api_sites__site_id__cloudflare_proxy_patch",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudflareProxyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Cloudflare Proxy",
        "tags": [
          "sites"
        ]
      }
    },
...
    "/api/sites/{site_id}/php": {
      "post": {
        "operationId": "change_php_version_api_sites__site_id__php_post",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePhpVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change Php Version",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/sites/{site_id}/php-settings": {
      "patch": {
        "operationId": "update_php_settings_api_sites__site_id__php_settings_patch",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhpSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Php Settings",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/sites/{site_id}/ssl": {
      "get": {
        "operationId": "site_ssl_status_api_sites__site_id__ssl_get",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Site Ssl Status",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/sites/{site_id}/ssl/renew": {
      "post": {
        "description": "Re-apply the Caddy config so it (re)attempts certificate issuance.\n\nCaddy renews valid certificates on its own; this endpoint covers the\nretry cases (initial issuance failed, DNS fixed after creation) and\nreturns a fresh probe of the resulting certificate status.",
        "operationId": "renew_site_ssl_api_sites__site_id__ssl_renew_post",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Renew Site Ssl",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/sites/{site_id}/wordpress": {
      "get": {
        "operationId": "wordpress_status_api_sites__site_id__wordpress_get",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WpStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Wordpress Status",
        "tags": [
          "sites"
        ]
      },
      "post": {
        "operationId": "install_wordpress_api_sites__site_id__wordpress_post",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WpInstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WpInstallAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Install Wordpress",
        "tags": [
          "sites"
        ]
      }
    },
...
  }
}